composer update

+++ app/careyshop/model/User.php
@@ -608,18 +608,18 @@ class User extends CareyShop
             return $this->setError('手机号码错误');
         }
 
-        //        // 验证验证码
-        //        $verifyDb = new Verification();
-        //        if (!$verifyDb->verVerification($data['mobile'], $data['code'])) {
-        //            return $this->setError($verifyDb->getError());
-        //        }
+        // 验证验证码
+        $verifyDb = new Verification();
+        if (!$verifyDb->verVerification($data['mobile'], $data['code'])) {
+            return $this->setError($verifyDb->getError());
+        }
 
         $result->setAttr('password', $data['password']);
         $result->save();
 
         Cache::tag('token:user_' . $result->getAttr('user_id'))->clear();
         $this->hasToken()->where(['client_id' => $result->getAttr('user_id'), 'client_type' => 0])->delete();
-        //$verifyDb->useVerificationItem(['number' => $data['mobile']]);
+        $verifyDb->useVerificationItem(['number' => $data['mobile']]);
 
         return true;
     }
